Ionic structure and nostr_private_key instead of noded up for robustness against some edge cases
