+++ src/State/store.ts
@@ -10,7 +10,7 @@ import historySlice, { storageKey as historyStorageKey, mergeLogic as historyMer
 import notificationSlice, { storageKey as notificationStorageKey, mergeLogic as notificationMergeLogic } from './Slices/notificationSlice';
 import generatedAssets from './Slices/generatedAssets';
 import loadingOverlay from './Slices/loadingOverlay';
-import nodedUpSlice from './Slices/nodedUpSlice';
+import nostrPrivateKey from './Slices/nostrPrivateKey';
 import subscriptionsSlice, { storageKey as subscriptionsStorageKey, mergeLogic as subscriptionsMergeLogic } from './Slices/subscriptionsSlice';
 import { useDispatch as originalUseDispatch, useSelector as originalUseSelector } from 'react-redux';
 import { MigrationFunction } from './Slices/migrations';
@@ -31,7 +31,7 @@ const store = configureStore({
     subscriptions: subscriptionsSlice,
     generatedAssets,
     loadingOverlay,
-    nodedUpSlice
+    nostrPrivateKey
   },
 });
 export type State = ReturnType<typeof store.getState>
